requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Pillow==10.0.1
python-dotenv==1.0.0
bleach==6.0.0
markdown==3.5.1
redis==5.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
boto3==1.28.62
sentry-sdk[flask]==1.32.0
python-sitemap==1.0.2
feedgen==0.9.0